requirements.txt
onnx
torch
openvino-dev
pyrealsense2